+++ src/commands/_common/types/CommandParser.ts
@@ -1,9 +1,11 @@
 import type { WritableDeep } from 'type-fest';
 import type { PipelineJson } from '../../../types/PipelineJson/PipelineJson';
 import type { TemplateJson } from '../../../types/PipelineJson/TemplateJson';
-import type { string_markdown_text } from '../../../types/typeAliases';
-import type { string_name } from '../../../types/typeAliases';
-import type { string_promptbook_documentation_url } from '../../../types/typeAliases';
+import type {
+    string_markdown_text,
+    string_name,
+    string_promptbook_documentation_url,
+} from '../../../types/typeAliases';
 import type { string_SCREAMING_CASE } from '../../../utils/normalization/normalizeTo_SCREAMING_CASE';
 import type { ___and___ } from '../../../utils/organization/___and___';
 import type { CommandUsagePlace } from './CommandUsagePlaces';
@@ -155,8 +157,8 @@ export type PipelineTemplateCommandParser<TCommand extends CommandBase> = Common
  * @private internal helper for command parsers
  */
 export type $TemplateJson = {
-    isBlockTypeSet: boolean;
-    isTemplateBlock: boolean;
+    isTemplateTypeSet: boolean;
+    isTemplate: boolean;
 } & Partial<WritableDeep<TemplateJson>>;
 //                         <- TODO: [🧠] `Partial<WritableDeep<...` vs `WritableDeep<Partial<...` - change ACRY