+++ src/types/PipelineJson/TemplateJsonCommon.ts
@@ -1,13 +1,15 @@
-import type { BlockType } from '../../commands/BLOCK/BlockTypes';
 import type { FormatCommand } from '../../commands/FORMAT/FormatCommand';
-import type { string_javascript } from '../typeAliases';
-import type { string_markdown } from '../typeAliases';
-import type { string_markdown_text } from '../typeAliases';
-import type { string_name } from '../typeAliases';
-import type { string_parameter_name } from '../typeAliases';
-import type { string_postprocessing_function_name } from '../typeAliases';
-import type { string_prompt } from '../typeAliases';
-import type { string_template } from '../typeAliases';
+import type { BlockType } from '../../commands/TEMPLATE/BlockTypes';
+import type {
+    string_javascript,
+    string_markdown,
+    string_markdown_text,
+    string_name,
+    string_parameter_name,
+    string_postprocessing_function_name,
+    string_prompt,
+    string_template,
+} from '../typeAliases';
 import type { Expectations } from './Expectations';
 
 /**